+++ src/main/java/app/weathertwin/repository/WeatherDataRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import app.weathertwin.entity.WeatherData;
 
-public interface WeatherDataRepository extends CrudRepository<WeatherData, Integer> {
+public interface WeatherDataRepository extends CrudRepository<WeatherData, Long> {
 
 }
